Modify Purchase Funcionality in ShoppingCar
Added a new method in BuyController to allow ShoppingCar.
Some errors related with prices in ShoppingCar stuff were fixed.

+++ ViewModels/ShoppingListViewModel.cs
@@ -9,14 +9,10 @@ namespace TrendyShop.ViewModels
     public class ShoppingListViewModel
     {
         public int Id { get; set; }
-
         public int ArticleId { get; set; }
         public Article Article { get; set; }
         public int Amount { get; set; }
-        public int Price { get; set; }
-
-        public int TotalPrice { get { return Amount * Price; } }
-
-
+        public double Price { get; set; }
+        public double TotalPrice { get { return Amount * Price; } }
     }
 }
